requirements.txt
torch~=2.8.0
torchvision~=0.23.0

# --- Vision & Graph Models ---
timm~=1.0.19
torch-geometric~=2.6.1

# --- 3D Processing & Utilities ---
# PyTorch3D is NOT listed here as it requires a special installation.
# See the installation notes below.
iopath               # PyTorch3D dependency
fvcore               # PyTorch3D dependency
ninja                # For building custom CUDA kernels
scikit-image         # For image processing and marching cubes
trimesh              # Robust library for loading/processing meshes
open3d               # 3D data processing and visualization

# --- Experimentation & Utilities ---
matplotlib           # For plotting
numpy
scipy
Pillow>=10.0         # Image loading
tensorboard          # For logging experiments (highly recommended)
tqdm~=4.67.1
notebook             # For running .ipynb files locally
pytorch3d~=0.7.8
dataclasses~=0.8